+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [unreleased] -->
 
+## [3.2.8] - 2025-03-11
+
+### Fix
+- Fix concept link and button usability in dark mode. 
+- Correct History Note description type (skosConcept)
+
 ## [3.2.7] - 2025-02-12
 
 ### Fix
@@ -307,7 +313,8 @@ is available.
 - Add changelog
 - Migrate SKOS Taxonomy Management schemas from local development branch to (this) plugin repository
 
-[unreleased]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.7...HEAD
+[unreleased]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.8...HEAD
+[3.2.8]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.7...v3.2.8
 [3.2.7]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.6...v3.2.7
 [3.2.6]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.5...v3.2.6
 [3.2.5]: https://github.com/andybywire/sanity-plugin-taxonomy-manager/compare/v3.2.4...v3.2.5